Improve the raw_size module.
- Added top level module documentation, specifies why it's there
and how should it be used, along with two documentation tests.
- Added two tests to check for an edge condition (the unreachable
block) and the roundtripping (writing, reading and compare the
results).
- Now uses `u64` instead of `usize`. `usize` is a platform dependent
type, it is pointer sized, so, in 8-bit,16-bit and 32-bit systems
a raw size integer may be too big to be held on an usize, `u64`
is guaranteed to work on all platforms and it also is the type
that can hold the maximum value that can be written.
- Added an error to check for the last point, so, when converting
from `u64` to `usize` and when it doesn't fit an error is returned.

Signed-off-by: Jean Pierre Dudey <me@jeandudey.tech>